Added VideoPreview component and timeline functionality
