+++ Vite-frontend/frontend/src/App.css
@@ -1,0 +1,220 @@
+/* Layout */
+.app-container {
+  display: flex;
+  height: 100vh;
+}
+
+/* Navbar */
+.navbar {
+  width: 160px;
+  height: 100vh;
+  background-color: #E6E6FA; /* Light Lavender */
+  color: #4B0082; /* Dark Lavender */
+  padding: 30px 20px;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 4px 0 10px rgba(75, 0, 130, 0.2);
+}
+
+.navbar h2 {
+  text-align: center;
+  margin-bottom: 25px;
+  font-size: 22px;
+}
+
+/* Navbar List */
+.navbar ul {
+  list-style: none;
+  padding: 0;
+  margin-top: 20px;
+}
+
+.navbar ul li {
+  padding: 12px;
+  text-align: left;
+}
+
+.navbar ul li a {
+  color: #4B0082;
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+  font-size: 18px;
+  font-weight: 500;
+  padding: 12px;
+  border-radius: 8px;
+  transition: 0.3s ease;
+}
+
+.icon {
+  margin-right: 12px;
+  font-size: 20px;
+}
+
+.navbar ul li a:hover {
+  background-color: #D8BFD8;
+  transform: scale(1.05);
+}
+
+/* Sidebar */
+.sidebar {
+  width: 220px;
+  height: 100vh;
+  background-color: #D8BFD8; /* Lighter Lavender */
+  color: #4B0082;
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  box-shadow: 2px 0 10px rgba(75, 0, 130, 0.1);
+}
+
+.upload-btn {
+  background-color: #9370DB; /* Medium Lavender */
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-size: 16px;
+  width: 100%;
+  margin-bottom: 20px;
+}
+
+.upload-btn:hover {
+  background-color: #7B68EE;
+}
+
+.sidebar ul {
+  list-style: none;
+  padding: 0;
+  width: 100%;
+}
+
+.sidebar li {
+  padding: 12px;
+  margin: 10px 0;
+  text-align: center;
+  cursor: pointer;
+  border-radius: 6px;
+  background-color: #E6E6FA;
+  transition: 0.3s;
+}
+
+.sidebar li:hover,
+.sidebar .active {
+  background-color: #C2A2DA;
+  color: white;
+}
+/* Video Container */
+.video-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  padding: 20px;
+}
+
+/* Video Box */
+.video-box {
+  background-color: #4B0082; /* Dark Lavender */
+  color: white;
+  width: 960px;
+  height: 540px;
+  border-radius: 12px;
+  box-shadow: 4px 4px 10px rgba(179, 43, 216, 0.2);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+}
+
+.video-title {
+  font-size: 20px;
+  margin-bottom: 20px;
+  margin-top: 60px;
+  color: #C2A2DA;
+}
+
+/* Video Player */
+.video-player {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 8px;
+}
+
+/* Video Placeholder */
+.video-placeholder {
+  font-size: 16px;
+  color: #ddd;
+  padding: 10px;
+  background-color: rgba(243, 175, 246, 0.1);
+  border-radius: 6px;
+}
+
+/* Timeline Container */
+.timeline-container {
+  position: relative;
+  width: 100%;
+  height: 120px; /* Increased height for more space */
+  background: #4B0082; /* Dark Lavender */
+  border-radius: 8px;
+  padding: 10px;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 2px 2px 8px rgba(75, 0, 130, 0.3);
+   /* Disable horizontal scrolling */
+}
+
+/* Waveform Container */
+.waveform-container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  height: 100px;
+  width: 100%;
+  background: transparent;
+  gap: 2rem;
+}
+
+/* Waveform */
+.waveform {
+  width: 100%;
+  height: 90px;
+}
+
+/* Play Button */
+.play-button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 60px;
+  height: 60px;
+  background: #efefef;
+  border-radius: 50%;
+  border: none;
+  outline: none;
+  cursor: pointer;
+  padding-bottom: 3px;
+  &:hover {
+    background: #ddd;
+  }
+}
+
+/* Timestamps */
+.timestamps {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  color: white;
+  font-size: 12px;
+  margin-top: 10px;
+}
+
+.timestamp {
+  flex: 1;
+  text-align: center;
+}
