--- frontend/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "subtitlegenerator",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
-    "@mui/material": "^6.4.3",
-    "@reduxjs/toolkit": "^2.5.1",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.7.9",
-    "react": "^19.0.0",
-    "react-contenteditable": "^3.3.7",
-    "react-dom": "^19.0.0",
-    "react-draggable": "^4.4.6",
-    "react-dropzone": "^14.3.5",
-    "react-icons": "^5.4.0",
-    "react-player": "^2.16.0",
-    "react-redux": "^9.2.0",
-    "react-router-dom": "^7.1.5",
-    "react-scripts": "^5.0.1",
-    "wavesurfer.js": "^7.9.0",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.20",
-    "postcss": "^8.5.1",
-    "tailwindcss": "^4.0.3"
-  }
-}
