--- frontend/src/components/PreviewTimeline.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Preview = () => {
-  return (
-    <div style={{ width: "70%", height: "300px", background: "black", margin: "auto" }}>
-      <h2 style={{ color: "white", textAlign: "center" }}>Video Preview</h2>
-    </div>
-  );
-};
-
-export default Preview;
